Add 100% test coverage to the config module and fix issues
Adds test coverage to the config module and modifies a few behaviors
based on it. ``pop`` calls result an exception, as keys should be
removed from the configuration. The default logging was still getting
set up early, so that's been removed so apps and the CLI can set it up
as they wish.

``load`` was placed in the LazyConfig class since there was no real
reason for it to be on its own.

``_validate`` was added to LazyConfig and is called at load time. If
validation fails, a ConfigurationException is raised. At this time it
asserts there are no extra keys in the configuration and enforces the
client properties (or resets them if the user specified just the app
name).

Signed-off-by: Jeremy Cline <jcline@redhat.com>