Merge _queues and _consumers in Twisted protocol
This combines the _queues set with the _consumers dictionary and uses a
new representation that includes the consumer tag, the queue for the
consumer, and the callback to be invoked when a message arrives. This
offers several advantages:

* Creating our own consumer tag rather than waiting for the server to
  create one for us, we can register consumers and return references to
  callers without actually calling out to the server. This will be used in
  the future with a ``consume`` and ``cancel`` API that doesn't require
  ``resumeProducing`` to be called first.

* Allows for Twisted users to have different callbacks for different
  consumers without starting one service per callback.

This is a prep-work commit and doesn't change the behavior of the
Twisted protocol.

Signed-off-by: Jeremy Cline <jcline@redhat.com>