Explicitly configure Travis notifications
The default behavior of Travis is to send notifications to the committer
and the author (but only if they have write access) when a build fails.
This no longer works well because mergify is the committer and the
author may not have write access.

Therefore, explicitly configure Travis to notify us via email when a
build fails. Additionally, it will notify the #fedora-apps channel on
freenode when the build first starts to fail.

Signed-off-by: Jeremy Cline <jcline@redhat.com>