Exclude the "sent-at" header from equality checks
In https://github.com/fedora-infra/bodhi/issues/2858 it was noted that
tests involving equality of messages fails because of the sent-at
header. It doesn't make a lot of sense to include this in an equality
check for the same reason the id is not included in the check - it's
metadata and of little interest to developers (unless they are tracking
messages).

Signed-off-by: Jeremy Cline <jcline@redhat.com>