Allow declaring --exchange/--queue-name/--routing-key independently
The command-line flags on the consume command are intended to make it
quick and easy to run simple consumers without fiddling with a
configuration file.

To this end, allow users to override --exchange, --queue-name, and
--routing-key, rather than requiring none or all. This brings it in line
with the manual page which states defaults for queue-name and exchange
will be used if they are not provided.

This implementation is simplistic in that it sets all bindings in the
configuration to be for the provided values. However, the command line
interface options do not offer a way exactly declare exchanges, queues,
and bindings (durability options, arbitrary arguments, etc) so the
documentation is updated to indicate for complex queue configurations,
the configuration file should be used instead of command-line arguments.

Signed-off-by: Jeremy Cline <jcline@redhat.com>