cli: Add a record sub-command
Add a new CLI command to record messages from a message broker and write
them to a file (or stdout). This was initially intended to be used to
collect test fixtures, but can also be chained together with other
commands using pipes. Messages are written as they arrive, one message
per line, so it's easy to, say, run a command when a message arrives:

$ fedora-messaging record - | while read -r line || -n $line;
    do echo "message received"; done

or create a quick (and completely unreliable) bridge between brokers:

$ fedora-messaging --conf=broker1.toml record - | \
     fedora-messaging --conf=broker2.toml publish

Signed-off-by: Sebastian Wojciechowski <s.wojciechowski89@gmail.com>
[Rebase and refactor]
Signed-off-by: Jeremy Cline <jcline@redhat.com>