Add Twisted consume APIs
This adds a new API to fedora-messaging, twisted_consume(), which
provides a way for users to start AMQP consumers that run until
explicitly canceled or until a fatal error occurs on the broker.

This introduces new versions of the Protocol, Factory, and Service
classes as the APIs of these Twisted classes had to change. The original
classes have been deprecated and will be removed in fedora-messaging
v2.0.0.

One thing to note is that these new classes use the standard library's
logging module. While it is true that these are blocking calls which may
use the network and happen in the reactor thread, it makes configuring
logging for consumers much easier.

Signed-off-by: Jeremy Cline <jcline@redhat.com>