Handle general connection/channel exceptions when publishing
If there's any channel or connection problem and we don't reset the
channel and connection, the next call to publish will fail with
something like a ChannelWrongStateError, so handle any ChannelError by
resetting everything.

This switches the exception handling order since NackError and company
are specific subclasses of AMQPChannelErrors and we handle those two
differently.

Signed-off-by: Jeremy Cline <jcline@redhat.com>