Improve the readability of the API documentation
This expands the message API documentation and adds a table of content
to the top of the API docs, separating each function or class with a
sub-header. It should provide a high level overview of the complete API
at a glance, and fleshes out some of the API documentation that was
lacking in the Message class.

Fixes: #164
Fixes: #166

Signed-off-by: Jeremy Cline <jcline@redhat.com>