Move all APIs to use the Twisted-managed connection
This introduces crochet[0], a small library designed to make it easy to
provide synchronous APIs using asynchronous Twisted APIs. This modifies
the `fedora_messaging.api.consume` and `fedora_messaging.api.publish`
calls to use the Twisted protocol.

There are several advantages here:

1. The synchronous and asynchronous APIs use the same code which means
   there's fewer opportunities for them to differ in behavior and less
   code to maintain.

2. Both APIs didn't really work well because they could not heartbeat in
   the background. This lets them do so because the connection is
   handled in Twisted in a separate crochet-managed thread. While the
   performance improvements are likely modest, it won't lead to so many
   TCP connections being killed by firewalls or the broker.

The downside is it introduces the complexity of threads to the API which
makes it less obvious what is going on.

There are a few minor changes here which slightly change the APIs:

1. Publishing now raises a PublishTimeout when the timeout is reached
   (30 seconds by default).

2. Previously, the Twisted consume API did not validate arguments like
   the synchronous version did, so it now raises a ValueError on invalid
   arguments instead of crashing in some undefined way.

3. Calling publish from the Twisted reactor thread now raises an
   exception instead of blocking the reactor thread.

Although all these are small changes, I think this should be in a 2.0
release. In practice I don't think any of these issues will bite anyone
so upgrading shouldn't be a problem for anyone.

This does not move the CLI back to the synchronous API (yet).

Signed-off-by: Jeremy Cline <jcline@redhat.com>