Update example config: extra properties, logging
This is based on my experience porting several fedmsg consumers
to fedora-messaging. @jcline thought it would be good for
consumers to provide more useful info in client_properties, so
I invented a couple of optional fields there. Also it seems
useful to note that if your consumer implements a logger, it
needs to be configured here or else messages won't show up.

Signed-off-by: Adam Williamson <awilliam@redhat.com>