Don't declare exchanges when consuming
The CLI (by way of Twisted) does not declare exchanges when consuming,
it assumes the exchange already exists. This seems reasonable, consumers
will still learn about configuration mistakes during the binding
process. This makes it impossible to consume from the public vhost in
Fedora, so this is a quick bugfix release while the transition to
Crochet and Twisted are worked on and polished.

Fixes: #171

Signed-off-by: Jeremy Cline <jcline@redhat.com>