Set the QoS on the channel that is created for the consumer
According to [the RabbitMQ
docs](https://www.rabbitmq.com/consumer-prefetch.html), they changed the
semantics of the `global` QoS parameter, and setting it to `True` as
we've done does not enforce the prefetch limit on all channels in the
connection but rather on all consumers in the channel. There is no way
to have connection-wide limits with RabbitMQ.

Since we create a new channel for each consumer, we must set the QoS in
this channel and not in the main channel, otherwise it'll always be
limitless.

Fixes #223

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>