cli: add a new "reconsume" command
This command is similar to the "replay" command, except it initializes a
consumer and passes it the message retrieved from datagrepper.

I recently implemented a new consumer and wrote myself a little CLI
which did this in order to test my consumer from an existing message.
It'd be nice for it to be available by default.

Signed-off-by: Jeremy Cline <jeremycline@microsoft.com>